Play with host test done.

+++ contracts/GameRockPaperScissors.sol
@@ -8,7 +8,7 @@ import "./Game.sol";
 contract RockPaperScissors is Game {
     // 石头剪刀布游戏
     // 选项: ROCK: 1; PAPER: 2; SCISSORS: 3;
-    constructor(uint8 _gameType, address _host, uint256 _wager) Game(_gameType, _host, _wager){
+    constructor(uint256 _gameType, address _host, uint256 _wager) Game(_gameType, _host, _wager){
     }
     
     function getWinnerAndLoser()
@@ -22,7 +22,7 @@ contract RockPaperScissors is Game {
     function getWinnerAndLoser(
         uint256[] memory _randomWords
     ) public override returns (address, address) {
-        uint8 result = uint8((_randomWords[0] % 3) + 1);
+        uint256 result = (_randomWords[0] % 3) + 1;
     
         if (isDefaultHost()) {
             Gambler memory playerHost = Gambler({id : host, choice : result});
@@ -62,7 +62,7 @@ contract RockPaperScissors is Game {
         // 游戏未完成，用户选择隐藏
         Gambler[] memory displayGamblers = new Gambler[](gamblers.length);
         if (isGameActive()) {
-            for (uint8 i = 0; i < gamblers.length; i++) {
+            for (uint256 i = 0; i < gamblers.length; i++) {
                     displayGamblers[i] = Gambler({
                     id : gamblers[i].id,
                     choice : DEFAULT_CHOICE
